Make sure the full reload starts before we evict all entries from the cache.
The values were inverted. The eviction of the cache happend 1h (after 24h) after the full reload would kick in (25h). The result of that would have been that during that 1h gap, the cache was not pre-warmed and messages that needed to be handled for devices, would have to go to query the twins to get the info back to the cache.