allow null values
If you could represent a null value in a twin collection, you would expect it to be read as null. It's impossible to do as null value props are removed in IoT Hub - see https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-device-twins#:~:text=Properties%20set%20to%20null%20are%20removed.

However, this case won't happen, it's still possible to repro it locally - you can create a TwinCollection with a null property which you then could not read with the implementation - well you could, and the end result would be the same (null) - but the method would give you an indication that it could not find the property. Now, it will indicate existance, however, in practice, when working against IoT Hub, it won't make a difference, as null and non existing is the same.