Basic version of standalone discovery service (#1507)
The LNS and the standalone discovery endpoint share
the same LNS discovery logic (for backwards compatibility). The standalone
discovery endpoint is designed that it can be deployed either on-prem
or in the cloud. We use tags to associate which LBS and LNS are in the same
network, which gives us the advantage of having dynamic discovery based
on only a matching tag. We do not integration test against a real IoT Hub
since that would further complicate the test setup.